% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bioabund.R
\name{calc_bioabund}
\alias{calc_bioabund}
\title{Calculate index of station-, district-, or region-level abundance and biomass}
\usage{
calc_bioabund(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = c("ALL", "BB", "NORTH", "NS", "PRIB", "STMATT", "UNSTRAT", "E166", "W166",
    "166TO173")[1],
  years = NULL,
  sex = NULL,
  size_min = NULL,
  size_max = NULL,
  crab_category = NULL,
  female_maturity = c("morphometric", "cutline")[1],
  shell_condition = NULL,
  egg_condition = NULL,
  clutch_size = NULL,
  bin_1mm = FALSE,
  spatial_level = c("stratum", "district", "region")[2],
  replace_retow = TRUE,
  rm_corners = FALSE
)
}
\arguments{
\item{crab_data}{Object created from \code{crabpack::get_specimen_data()}.}

\item{species}{Character string. One of \code{c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR")}.}

\item{region}{Character string describing the region of interest. One of
\verb{c("EBS", "NBS"}). Defaults to \code{"EBS"} for Eastern Bering Sea.}

\item{district}{Character string. One or many of \code{c("ALL", "BB", "NORTH", "NS", "PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173")}.
Defaults to \code{"ALL"} districts within the selected region if not specified.}

\item{years}{Numeric or integer vector of years. Default output is all available years in data.}

\item{sex}{Character string. One or both of \code{c("male", "female")}. Optional,
\code{"male"} or \code{"female"} only will provide estimates for the selected sex,
specifying both \code{"male"} and \code{"female"} will provide estimates for
each of the selected sexes.}

\item{size_min}{Integer. Optional, desired lower range of crab size (inclusive).}

\item{size_max}{Integer. Optional, desired upper range of crab size (inclusive).}

\item{crab_category}{Character string. One or many of
\code{c("mature_male", "large_male", "legal_male", "preferred_male", "immature_male", "small_male", "sublegal_male", "mature_female", "immature_female", "all_categories")}.
Optional, specifying this parameter will provide estimates
for each of the selected categories; \code{"all_categories"} will
provide estimates for each relevant category for the given species.
If using a female category, maturity will be based on morphometric
maturity (default \code{female_maturity = "morphometric"}). Set
\code{female_maturity = "cutline"} if you want to define female maturity
based on ADF&G size cutlines. See \code{crabpack::set_crab_category()}
documentation for description of species-specific category combinations.}

\item{female_maturity}{Character string. One of \code{c("morphometric", "cutline")}.
Defaults to \code{"morphometric"} maturity for female crab. Morphometric
maturity biomass and abundance estimates are not available for
male crab at this time.}

\item{shell_condition}{Character string. One or many of \code{c("soft_molting", "new_hardshell", "oldshell", "very_oldshell", "all_categories")}.
Optional, specifying this parameter will provide estimates for
each of the selected shell conditions; \code{"all_categories"} will
provide estimates for each available shell condition category.}

\item{egg_condition}{Character string. One or many of \code{c("none", "uneyed", "eyed", "dead", "empty_cases", "hatching", "all_categories")}.
Optional, specifying this parameter will provide estimates for each of the
selected egg conditions; \code{"all_categories"} will provide estimates
for each available egg condition category. Note that specifying
\code{egg_condition} will return only female specimens in the final output.}

\item{clutch_size}{Character string. One or many of \code{c("immature", "mature_barren", "trace", "quarter", "half", "three_quarter", "full", "all_categories")}.
Optional, specifying this parameter will provide estimates for each of
the selected clutch sizes; \code{"all_categories"} will provide estimates for
each available clutch size category. Note that specifying \code{clutch_size}
will return only female specimens in the final output.}

\item{bin_1mm}{Boolean T/F. If \code{TRUE}, estimates will be provided for each 1mm bin
within the size range specified in \code{size_min} and/or \code{size_max},
or for the full range of observed sizes in the data. Defaults to \code{FALSE}.}

\item{spatial_level}{Character string. One of \code{c("stratum", "district", "region")}.
Describes the spatial resolution of biomass and abundance output,
aggregated to the stratum, district (default), or regional scale.}

\item{replace_retow}{Boolean T/F. If \code{TRUE}, replace female Bristol Bay red king crab observations
with resampled data (haul type = 17) in years when a Bristol Bay
retow took place. Defaults to \code{TRUE}, please use care when interpreting
BBRKC outputs if using \code{FALSE}.}

\item{rm_corners}{Boolean T/F. If \code{TRUE}, remove corner stations from Pribilof and St. Matthew
Districts and calculate CPUE, abundance, and biomass based on a single
stratum within the surrounding district. Defaults to \code{FALSE}, this is to
be used for design-based data exercises only.}
}
\value{
A data frame with area swept-expanded estimates of stratum-, district-, or
region-level abundance, biomass (mt), and biomass (lbs) by year.
\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
ABUNDANCE          \tab Estimated crab abundance using an area-swept expansion.                                                                                               \cr
ABUNDANCE_CI       \tab Estimated crab abundance 95\% confidence interval.                                                                                                     \cr
ABUNDANCE_CV       \tab Estimated crab abundance coefficient of variation.                                                                                                    \cr
BIOMASS_LBS        \tab Estimated crab biomass in pounds.                                                                                                                     \cr
BIOMASS_LBS_CI     \tab Estimated crab biomass (pounds) 95\% confidence interval.                                                                                              \cr
BIOMASS_LBS_CV     \tab Estimated crab biomass (pounds) coefficient of variation.                                                                                             \cr
BIOMASS_MT         \tab Estimated crab biomass in metric tons.                                                                                                                \cr
BIOMASS_MT_CI      \tab Estimated crab biomass (metric tons) 95\% confidence interval.                                                                                         \cr
BIOMASS_MT_CV      \tab Estimated crab biomass (metric tons) coefficient of variation.                                                                                        \cr
CATEGORY           \tab Shorthand name of crab sex-size-maturity category.                                                                                                    \cr
CLUTCH_TEXT        \tab Text describing  amount of eggs, if any, in female crabs.                                                                                             \cr
DISTRICT           \tab Name of crab district, shorthand.                                                                                                                     \cr
EGG_CONDITION_TEXT \tab Text describing egg condition category in female crabs.                                                                                               \cr
REGION             \tab Survey region. EBS = Eastern Bering Sea, NBS = Northern Bering Sea.                                                                                   \cr
SEX_TEXT           \tab Text describing sex of crab.                                                                                                                          \cr
SHELL_TEXT         \tab Text describing carapace condition of crab.                                                                                                           \cr
SIZE_1MM           \tab Measurement of crab in 1 millimeter bins. This number is carapace length for king crab and hair crab species, and carapace width for Chionoecetes spp.\cr
SPECIES            \tab Name of species, shorthand.                                                                                                                           \cr
TOTAL_AREA         \tab Total area of stratum, in square nautical miles.                                                                                                      \cr
YEAR               \tab Year the observation (survey) was collected.
}
}
\description{
This function calculates indices of abundance and biomass (mt and lbs)
at the station-, district-, or region-level for a given crab species.
Optional arguments also allow these indices to be calculated for subsets of
crab by biometric categories such as size, sex, maturity, and shell condition.
}
