% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_male_ogives.R
\name{get_male_ogives}
\alias{get_male_ogives}
\title{Pull time series of estimated proportion mature by 10mm bin for Chionoecetes males}
\usage{
get_male_ogives(
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = c("ALL", "E166", "W166")[1],
  output = c("estimates", "parameters")[1],
  channel = NULL
)
}
\arguments{
\item{species}{character string. One of c("Tanner Crab", "Snow Crab").
c("TANNER", "SNOW").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One of c("ALL", "E166", "W166"). Defaults to
"ALL", "E166" and "W166" are to be used for Tanner Crab only.}

\item{output}{character string. One of c("estimates", "parameters"). Defaults
to "estimates" to provide ___. Specify "parameters" for a table
of model parameter estimates for each year.}

\item{channel}{connection to Oracle created via crabPack::get_connected() or RODBC::odbcConnect().
'channel = "local"' is to be used for internal development purposes only.}
}
\value{
TBD
}
\description{
.... time series of Tanner Crab and Snow Crab male ogives...proportion mature
Also yearly parameter estimates....
}
