% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maturity.R
\name{get_maturity}
\alias{get_maturity}
\title{PURPOSE: assign maturity-sex categories based on stock-specific size definitions
ARGUMENTS: crab_dat, stock
#' NEEDS MAT_LOOKUP!}
\usage{
get_maturity(crab_dat, stock)
}
\arguments{
\item{crab_dat}{Species-specific CrabHaul file}

\item{stock}{Character
"BBRKC", "PribRKC", "NSRKC", "NorthRKC", "PribBKC", "StMattBKC", "BKCNBS",
"TannerE", "TannerW", "TannerWNBS", "Snow", "SnowNBS",
"Hybrid", "HybridNBS", "Hair", "HairNBS"}
}
\value{
Returns a data frame with catch data and a MAT_SEX column with stock-specific maturity designations
}
\description{
PURPOSE: assign maturity-sex categories based on stock-specific size definitions
ARGUMENTS: crab_dat, stock
#' NEEDS MAT_LOOKUP!
}
