% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimen_data.R
\name{get_specimen_data}
\alias{get_specimen_data}
\title{Pull AFSC Shellfish Assessment Program Bering Sea survey data}
\usage{
get_specimen_data(
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  years = NULL,
  channel = NULL
)
}
\arguments{
\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One or many of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{years}{numeric or integer vector of years.}

\item{channel}{connection to Oracle created via \code{crabpack::get_connected()} or \code{RODBC::odbcConnect()}.}
}
\value{
a named list containing specimen, haul, stratum, area, and size group
information for the region, districts, years, and species of interest.
}
\description{
Pulls haul, specimen, and stratum grouping information for the
region, districts, years, and species of interest from the
CRABBASE schema in the AKFIN Oracle database.
}
