% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_crab_category.R
\name{set_crab_category}
\alias{set_crab_category}
\title{Assign size-sex-maturity categories based on species- and district-specific size definitions}
\usage{
set_crab_category(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  crab_category = NULL,
  female_maturity = c("morphometric", "cutline")[1]
)
}
\arguments{
\item{crab_data}{object created from \code{crabpack::get_data()}.}

\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One or many of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{crab_category}{character string. One or many of c("legal_male", "preferred_male",
"mature_male", "immature_male", "mature_female", "immature_female",
"all_categories"). Optional, specifying this parameter will provide
estimates for each of the selected categories; "all_categories" will
provide estimates for each relevant category for the given species.
If using a female category, maturity will be based on morphometric
maturity (default "morphometric" for the optional \code{female_maturity}
parameter). Set \code{female_maturity} to "cutline" if you want to define
female maturity based on ADF&G size cutlines.}

\item{female_maturity}{character string. One of c("morphometric", "cutline").
Defaults to "morphometric" maturity for female crab. Morphometric
maturity biomass and abundance estimates are not available for
male crab at this time.}
}
\value{
a data frame with crab specimen data and an additional CATEGORY column
with stock-specific maturity designations.
}
\description{
TBD
}
