% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bioabund.R
\name{calc_bioabund}
\alias{calc_bioabund}
\title{Calculate index of station-, district-, or region-level abundance and biomass}
\usage{
calc_bioabund(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = c("ALL", "BB", "NORTH", "NS", "PRIB", "STMATT", "UNSTRAT", "E166", "W166",
    "166TO173")[1],
  years = NULL,
  sex = NULL,
  size_min = NULL,
  size_max = NULL,
  crab_category = NULL,
  female_maturity = c("morphological", "cutline")[1],
  shell_condition = NULL,
  egg_condition = NULL,
  clutch_size = NULL,
  bin_1mm = FALSE,
  spatial_level = c("station", "district", "region")[2],
  replace_retow = TRUE
)
}
\arguments{
\item{crab_data}{object created from \code{crabPack::get_data()}.}

\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{years}{numeric or integer vector of years.}

\item{sex}{character string. One or both of c("male", "female"). Optional,
"male" or "female" only will provide estimates for the selected sex,
specifying both "male" and "female" will provide estimates for
each of the selected sexes.}

\item{size_min}{integer. Optional, desired lower range of crab sizes (inclusive).}

\item{size_max}{integer. Optional, desired upper range of crab sizes (inclusive).}

\item{crab_category}{character string. One of c("legal_male", "preferred_male", "mature_male",
"immature_male", "mature_female", "immature_female", "all").
Optional, specifying this parameter will provide estimates for
each of the selected categories; "all" will provide estimates for each
relevant category for the given species. If using a female category,
maturity will be based on morphological maturity (default "morphological"
for the optional 'female_maturity' parameter). Set 'female_maturity'
to "cutline" if you want to define female maturity based on ADF&G size cutlines.}

\item{female_maturity}{character string. One of c("morphological", "cutline").
Defaults to "morphological" maturity for female crab. Morphological
maturity designation for male crab are not available at this time.}

\item{shell_condition}{character string. One or many of c("soft molting",
"new hardshell", "oldshell", "very oldshell", "all_categories").
Optional, specifying this parameter will provide estimates
for each of the selected shell conditions; "all_categories" will
provide estimates for each available shell condition category.}

\item{egg_condition}{character string. One or many of c("none", "uneyed", "eyed",
"dead", "empty cases", "hatching", "all_categories"). Optional,
specifying this parameter will provide estimates for each of the
selected egg conditions; "all_categories" will provide estimates
for each available egg condition category.}

\item{clutch_size}{character string. One or many of c("immature", "mature barren",
"trace", "quarter", "half", "three quarter", "full", "all_categories").
Optional, specifying this parameter will provide estimates for each of
the selected clutch sizes; "all_categories" will provide estimates for
each available clutch size category.}

\item{bin_1mm}{boolean T/F. If TRUE, estimates will be provided for each 1mm bin
within the size range specified in 'size_min' and/or 'size_max',
or for the full range of observed sizes in the data. Defaults to FALSE.}

\item{spatial_level}{character string. One of c("station", "district", "region").
Describes the spatial resolution of biomass and abundance output,
either kept at the haul level, or aggregated to the district (default)
or regional scale.}

\item{replace_retow}{boolean T/F. Replace female Bristol Bay red king crab observations
with resampled data (haul type "17") in years when a Bristol Bay
retow took place? Defaults to TRUE, please use care when interpreting
BBRKC outputs if using FALSE.}
}
\value{
A data frame with area swept-expanded estimates of species- or district-level abundance,
biomass (mt), and biomass (lbs) by year.
}
\description{
This function calculates indices of abundance and biomass (MT and LBS)
at the station-, district-, or region-level for a given crab species.
Optional arguments also allow these indices to be calculated for subsets of
crab by biometric categories such as size, sex, maturity, and shell condition.
}
