% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cpue.R
\name{calc_cpue}
\alias{calc_cpue}
\title{Calculate haul-level catch per unit effort}
\usage{
calc_cpue(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = c("ALL", "BB", "NORTH", "NS", "PRIB", "STMATT", "UNSTRAT", "E166", "W166",
    "166TO173")[1],
  years = NULL,
  sex = NULL,
  size_min = NULL,
  size_max = NULL,
  crab_category = NULL,
  female_maturity = c("morphometric", "cutline")[1],
  shell_condition = NULL,
  egg_condition = NULL,
  clutch_size = NULL,
  bin_1mm = FALSE,
  replace_retow = TRUE,
  output = c("cpue", "bioabund")[1]
)
}
\arguments{
\item{crab_data}{object created from \code{crabpack::get_data()}.}

\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One or many of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{years}{numeric or integer vector of years.}

\item{sex}{character string. One or both of c("male", "female"). Optional,
"male" or "female" only will provide estimates for the selected sex,
specifying both "male" and "female" will provide estimates for
each of the selected sexes.}

\item{size_min}{integer. Optional, desired lower range of crab size (inclusive).}

\item{size_max}{integer. Optional, desired upper range of crab size (inclusive).}

\item{crab_category}{character string. One or many of c("legal_male", "preferred_male",
"mature_male", "immature_male", "mature_female", "immature_female",
"all_categories"). Optional, specifying this parameter will provide
estimates for each of the selected categories; "all_categories" will
provide estimates for each relevant category for the given species.
If using a female category, maturity will be based on morphometric
maturity (default "morphometric" for the optional \code{female_maturity}
parameter). Set \code{female_maturity} to "cutline" if you want to define
female maturity based on ADF&G size cutlines.}

\item{female_maturity}{character string. One of c("morphometric", "cutline").
Defaults to "morphometric" maturity for female crab. Morphometric
maturity biomass and abundance estimares for male crab are not
available at this time.}

\item{shell_condition}{character string. One or many of c("soft molting", "new hardshell",
"oldshell", "very oldshell", "all_categories"). Optional,
specifying this parameter will provide estimates for each of the
selected shell conditions; "all_categories" will provide estimates
for each available shell condition category.}

\item{egg_condition}{character string. One or many of c("none", "uneyed", "eyed",
"dead", "empty cases", "hatching", "all_categories"). Optional,
specifying this parameter will provide estimates for each of the
selected egg conditions; "all_categories" will provide estimates
for each available egg condition category. Note that specifying
\code{egg_condition} will return only female specimens in the final output.}

\item{clutch_size}{character string. One or many of c("immature", "mature barren",
"trace", "quarter", "half", "three quarter", "full", "all_categories").
Optional, specifying this parameter will provide estimates for each of
the selected clutch sizes; "all_categories" will provide estimates for
each available clutch size category. Note that specifying \code{clutch_size}
will return only female specimens in the final output.}

\item{bin_1mm}{boolean T/F. If TRUE, estimates will be provided for each 1mm bin
within the size range specified in 'size_min' and/or 'size_max',
or for the full range of observed sizes in the data. Defaults to FALSE.}

\item{replace_retow}{boolean T/F. Replace female Bristol Bay red king crab observations
with resampled data (haul type "17") in years when a Bristol Bay
retow took place? Defaults to TRUE, please use care when interpreting
BBRKC outputs if using FALSE.}

\item{output}{character string. One of c("cpue", "bioabund"). Default \code{output}
is "cpue"; "bioabund" is used for internal package purposes only, as
\code{calc_cpue()} is called within \code{calc_bioabund()}.}
}
\value{
a data frame with station-level crab counts and area swept-expanded
estimates of weight (mt, lbs) and numerical CPUE by year.
}
\description{
This function calculates and zero-fills weight (mt, lbs) and numerical
catch per unit effort (area swept, nm2).
}
