% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crabPack-package.R
\docType{package}
\name{crabpack-package}
\alias{crabpack}
\alias{crabpack-package}
\title{crabpack: Calculate Bering Sea Crab CPUE, Abundance, and Biomass}
\description{
This package contains functions to calculate standard NOAA AFSC Shellfish Assessment Program summer bottom trawl survey data products (CPUE, Abundance, and Biomass) for the Eastern Bering Sea Shelf (1975-present) and Northern Bering Sea Shelf (2010-present).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://afsc-shellfish-assessment-program.github.io/crabpack/}
}

}
\author{
\strong{Maintainer}: Shannon Hennessey \email{shannon.hennessey@noaa.gov}

}
\keyword{internal}
