% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimen_data.R
\name{get_specimen_data}
\alias{get_specimen_data}
\title{Pull AFSC Shellfish Assessment Program Bering Sea survey data}
\usage{
get_specimen_data(
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  years = NULL,
  channel = NULL
)
}
\arguments{
\item{species}{Character string. One of \code{c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR")}.}

\item{region}{Character string describing the region of interest. One of
\verb{c("EBS", "NBS"}). Defaults to \code{"EBS"} for Eastern Bering Sea.}

\item{district}{Character string. One or many of \code{c("ALL", "BB", "NORTH", "NS", "PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173")}.
Defaults to \code{"ALL"} districts within the selected region if not specified.}

\item{years}{Numeric or integer vector of years.}

\item{channel}{Connection to Oracle created via \code{crabpack::get_connected()} or \code{RODBC::odbcConnect()}.}
}
\value{
A named list containing specimen, haul, stratum, area, and size group
information for the region, districts, years, and species of interest.
}
\description{
Pulls haul, specimen, and stratum grouping information for the
region, districts, years, and species of interest from the
CRABBASE schema in the AKFIN Oracle database.
}
