% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specimen_data.R
\name{get_specimen_data}
\alias{get_specimen_data}
\title{Pull AFSC Shellfish Assessment Program Bering Sea survey data}
\usage{
get_specimen_data(
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  years = NULL,
  channel = NULL
)
}
\arguments{
\item{species}{character string. One of c("Red King Crab", "Blue King Crab",
"Tanner Crab", "Snow Crab", "Hybrid Tanner Crab", "Horsehair Crab").
c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{years}{numeric or integer vector of years.}

\item{channel}{connection to Oracle created via crabPack::get_connected() or RODBC::odbcConnect().
'channel = "local"' is to be used for internal development purposes only.}
}
\value{
a named list containing survey, cruise, haul, catch, size, specimen,
stratum, subarea, and stratum_groups information for the years and species of interest.
}
\description{
TBD
}
