% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_crab_category.R
\name{set_crab_category}
\alias{set_crab_category}
\title{Assign maturity-sex categories based on stock-specific size definitions
#' NEEDS MAT_LOOKUP!}
\usage{
set_crab_category(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  crab_category = NULL,
  female_maturity = c("morphological", "cutline")[1]
)
}
\arguments{
\item{crab_data}{object created from \code{crabpack::get_data()}.}

\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{region}{character string describing the region of interest. One of
c("EBS", "NBS"). Defaults to "EBS" for Eastern Bering Sea.}

\item{district}{character string. One of c("ALL", "BB", "NORTH", "NS",
"PRIB", "STMATT", "UNSTRAT", "E166", "W166", "166TO173").
Defaults to "ALL" districts within the selected region if not specified.}

\item{crab_category}{character string. One of c("legal_male", "preferred_male", "mature_male",
"immature_male", "mature_female", "immature_female", "all").
Optional, specifying this parameter will provide estimates for
each of the selected categories; "all" will provide estimates for each
relevant category for the given species. If using a female category,
maturity will be based on morphological maturity (default "morphological"
for the optional 'female_maturity' parameter). Set 'female_maturity'
to "cutline" if you want to define female maturity based on ADF&G size cutlines.}

\item{female_maturity}{character string. One of c("morphological", "cutline").
Defaults to "morphological" maturity for female crab. Morphological
maturity designation for male crab are not available at this time.}
}
\value{
Returns a data frame with crab specimen data and a CATEGORY column with
stock-specific maturity designations
}
\description{
TBD
}
