% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_crab_category.R
\name{set_crab_category}
\alias{set_crab_category}
\title{Assign size-sex-maturity categories based on species- and district-specific size definitions}
\usage{
set_crab_category(
  crab_data = NULL,
  species = NULL,
  region = c("EBS", "NBS")[1],
  district = NULL,
  crab_category = NULL,
  female_maturity = c("morphometric", "cutline")[1]
)
}
\arguments{
\item{crab_category}{Character string. One or many of
\code{c("mature_male", "large_male", "legal_male", "preferred_male", "immature_male", "small_male", "sublegal_male", "mature_female", "immature_female", "all_categories")}.
Optional, specifying this parameter will provide estimates
for each of the selected categories; \code{"all_categories"} will
provide estimates for each relevant category for the given species.
If using a female category, maturity will be based on morphometric
maturity (default \code{female_maturity = "morphometric"}). Set
\code{female_maturity = "cutline"} if you want to define female maturity
based on ADF&G size cutlines.}
}
\value{
A data frame with crab specimen data and an additional "CATEGORY" column
with stock-specific maturity designations.
}
\description{
A utility function to assign standard size-sex-maturity categories to crab
specimen data. Standard categories for Red King Crab and Blue King Crab are
\code{immature_male}, \code{mature_male}, \code{legal_male}, \code{immature_female}, and \code{mature_female}.
Categories for Tanner Crab, Snow Crab, and Hybrid Crab are \code{small_male}, \code{large_male},
\code{legal_male}, \code{preferred_male}, \code{immature_female}, and \code{mature_female}. Categories
for Hair Crab are \code{sublegal_male}, \code{legal_male}, and \code{female}.
}
