% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_variables.R
\name{set_variables}
\alias{set_variables}
\title{Set optional variables used in \code{calc_cpue} and \code{calc_bioabund} and subset data}
\usage{
set_variables(
  crab_data = NULL,
  species = NULL,
  sex = NULL,
  size_min = NULL,
  size_max = NULL,
  crab_category = NULL,
  female_maturity = c("morphological", "cutline")[1],
  shell_condition = NULL,
  egg_condition = NULL,
  clutch_size = NULL,
  bin_1mm = NULL
)
}
\arguments{
\item{crab_data}{object created from \code{crabPack::get_data()}.}

\item{species}{character string. One of c("RKC", "BKC", "TANNER", "SNOW", "HYBRID", "HAIR").}

\item{sex}{character string. One or both of c("male", "female"). Optional,
"male" or "female" only will provide estimates for the selected sex,
specifying both "male" and "female" will provide estimates for
each of the selected sexes.}

\item{size_min}{integer. Optional, desired lower range of crab sizes (inclusive).}

\item{size_max}{integer. Optional, desired upper range of crab sizes (inclusive).}

\item{crab_category}{character string. One of c("legal_male", "preferred_male", "mature_male",
"immature_male", "mature_female", "immature_female", "all").
Optional, specifying this parameter will provide estimates for
each of the selected categories; "all" will provide estimates for each
relevant category for the given species. If using a female category,
maturity will be based on morphological maturity (default "morphological"
for the optional 'female_maturity' parameter). Set 'female_maturity'
to "cutline" if you want to define female maturity based on ADF&G size cutlines.}

\item{female_maturity}{character string. One of c("morphological", "cutline").
Defaults to "morphological" maturity for female crab. Morphological
maturity designation for male crab are not available at this time.}

\item{shell_condition}{character string. One or many of c("soft molting",
"new hardshell", "oldshell", "very oldshell", "all_categories").
Optional, specifying this parameter will provide estimates
for each of the selected shell conditions; "all_categories" will
provide estimates for each available shell condition category.}

\item{egg_condition}{character string. One or many of c("none", "uneyed", "eyed",
"dead", "empty cases", "hatching", "all_categories"). Optional,
specifying this parameter will provide estimates for each of the
selected egg conditions; "all_categories" will provide estimates
for each available egg condition category.}

\item{clutch_size}{character string. One or many of c("immature", "mature barren",
"trace", "quarter", "half", "three quarter", "full", "all_categories").
Optional, specifying this parameter will provide estimates for each of
the selected clutch sizes; "all_categories" will provide estimates for
each available clutch size category.}

\item{bin_1mm}{boolean T/F. If TRUE, estimates will be provided for each 1mm bin
within the size range specified in 'size_min' and/or 'size_max',
or for the full range of observed sizes in the data. Defaults to FALSE.}
}
\value{
....
}
\description{
A utility function to...
}
